pyproject.toml
[tool.poetry]
name = "pygame-tetris"
version = "0.1.0"
description = "Project work for the course Ohjelmistotekniikka"
authors = ["Antti Perttula <anttiperttula@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.6"
pygame = "^2.0.1"
invoke = "^1.5.0"
python-dotenv = "^0.17.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
coverage = "^5.5"
pylint = "^2.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
